+++ TeamCode/build.gradle
@@ -37,6 +37,5 @@ dependencies {
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.12.7'
     implementation 'com.acmerobotics.roadrunner:core:0.5.6'
     implementation "com.acmerobotics.dashboard:dashboard:0.4.16"
-    implementation 'com.qualcomm.ftcrobotcontroller:RobotCore:8.0.0' // Or latest version
 
 }
